requirements.txt
-e .
aws-cdk.core
aws-cdk.aws-ec2
aws-cdk.aws-iam
aws-cdk.aws-eks
